+++ src/config/common.php
@@ -80,7 +80,7 @@ return [
     'params' => [
         'adminEmail' => getenv('APP_ADMIN_EMAIL'),
         'context.menuItems' => [],
-        'backend.iframe.name' => '_self',
+        'backend.iframe.name' => getenv('APP_PARAMS_BACKEND_IFRAME_NAME') ?: '_self',
     ],
     'components' => [
         'assetManager' => [